backend/requirements.txt
Flask
python-dotenv
pyjwt